edit
If we get inside this "if", this means that pid file exists;
but it can only exist if it is not the first "showit" run, because
otherwise the pid file would have been removed in initscreen().
And as such, the second "if" is redundant. But this change was not tested yet,
so beware...