fix bison error on line 375
pascal.y has this:

constants:
         | PCONST constdefinitions
         | PCONST constdefinitions conststringdefinition
         ;

Line 375 is in section "@<Constants...@>=", which is used as

    const @<Constants in the outer block@>@/

So, to match the grammar, constringdefinition must be in the end of this section.
In tex.web it is pool_name. But in mf.web there are non-strings after pool_name.
So we move pool_name to the end of section in mf.web and the error goes away.